Scripts: compile miui dtbo after compilation
This is done through a new patch file that gets applied after the
compilation process is complete in what I hope will just generate a
brand new DTBO file to package for us, we will also clone and checkout
to a much older revision of our AK3 script

This also reverts commit 2b0dfc95c91e67435e6dc56f2b0cf33f185371e6

+++ build.sh
@@ -5,7 +5,7 @@
 
 SECONDS=0 # builtin bash timer
 
-ZIPNAME="[OSS]STRIX-sweet-revival-$(date '+%Y%m%d-%H%M').zip"
+ZIPNAME="STRIX-sweet-revival-$(date '+%Y%m%d-%H%M').zip"
 
 export ARCH=arm64
 export KBUILD_BUILD_USER=vbajs
@@ -47,12 +47,34 @@ if [ ! -f "$kernel" ] || [ ! -f "$dtbo" ] || [ ! -f "$dtb" ]; then
 	exit 1
 fi
 
+echo -e "\nCompiled Kernel + OSS dimensions, now compiling MIUI dimensions while dirty.."
+mkdir ./out/arch/arm64/boot/oss/
+cp $dtbo out/arch/arm64/boot/oss/dtbo.img
+ossdtbo="out/arch/arm64/boot/oss/dtbo.img"
+rm -rf $dtbo
+patch -p1 < miui-dtbo.patch
+make -j$(nproc --all) \
+    O=out \
+    ARCH=arm64 \
+    LLVM=1 \
+    LLVM_IAS=1 \
+    CROSS_COMPILE=aarch64-linux-gnu- \
+    CROSS_COMPILE_COMPAT=arm-linux-gnueabi-
+
+if [ ! -f "$dtbo" ]; then
+	echo -e "\nCompilation failed!"
+	exit 1
+fi
+
 echo -e "\nKernel compiled successfully! Zipping up...\n"
+mkdir ./out/arch/arm64/boot/miui/
+cp $dtbo out/arch/arm64/boot/miui/dtbo.img
+miuidtbo="out/arch/arm64/boot/miui/dtbo.img"
 
 if [ -d "$AK3_DIR" ]; then
 	cp -r $AK3_DIR AnyKernel3
 else
-	if ! git clone -q --depth=1 https://github.com/vbajs/AnyKernel3.git -b fiqri AnyKernel3; then
+	if ! git clone -q https://github.com/vbajs/AnyKernel3.git -b fiqri AnyKernel3; then
 		echo -e "\nAnyKernel3 repo not found locally and couldn't clone from GitHub! Aborting..."
 		exit 1
 	fi
@@ -62,13 +84,18 @@ fi
 sed -i "s/device\.name1=.*/device.name1=sweet/" AnyKernel3/anykernel.sh
 sed -i "s/device\.name2=.*/device.name2=sweetin/" AnyKernel3/anykernel.sh
 
+cd AnyKernel3
+git reset --hard a74cbbf53f97245f6441663cfa0bc58db39aee53
+cd ..
 cp $kernel AnyKernel3
-cp $dtbo AnyKernel3
+cp $ossdtbo AnyKernel3/dtbo/oss
+cp $miuidtbo AnyKernel3/dtbo/miui
 cp $dtb AnyKernel3
 cd AnyKernel3
-zip -r9 "../$ZIPNAME" * -x .git
+zip -r9 "../$ZIPNAME" * -x .git README.md
 cd ..
 rm -rf AnyKernel3
+git restore arch/arm64/boot/dts/qcom/dsi-panel-k6-38-0c-0a-fhd-dsc-video.dtsi
 echo -e "\nCompleted in $((SECONDS / 60)) minute(s) and $((SECONDS % 60)) second(s) !"
 echo "Zip: $ZIPNAME"
